requirements/requirements-testing.txt
pluggy==0.3.1
psycopg2==2.6.1
py==1.4.31
pytest==2.8.7
pytest-django==2.9.1
tox==2.3.1
